requirements.txt
requests>=2.31.0

# API clients for AI judges
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.8.0
